* CRITICAL * Updated yii2 version to 2.0.10
This update is needed to support aliases in the optional param in AuthMethod class of yii2. This is needed to configure the public api to allow guest and logged users with no errors.
If this version update needs to be reverted, you need to find other solution to the problem with public api, that need to know the connected user in some situations (for example, to determine if a product is loved by current user or not)